Type-Punning

+++ base-hack/include/dk64.h
@@ -248,7 +248,7 @@ extern void save(void);
 extern void* getSpawnerTiedActor(short target_trigger, short props_change);
 
 extern void _guScaleF(void* mtx, int x, int y, int z);
-extern void _guTranslateF(void* mtx, int x, int y, int z);
+extern void _guTranslateF(void* mtx, float x, float y, float z);
 extern void _guMtxCatF(void* mtx, void* unk0, void* unk1);
 extern void _guMtxF2L(void* mtx, void* unk0);
 extern void _guMtxXFML(void* unk0, int unk1, int unk2, int unk3, float* x, float* y, float* z);
